AzureFunctions-Python-api/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
attrs==24.3.0
azure-common==1.1.28
azure-core==1.32.0
azure-data-tables>=12.4.0
azure-functions==1.21.3
azure-functions-durable>=1.2.0
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
requests
bcrypt>=4.0.1
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.0
distro==1.9.0
Django==5.1.5
django-appconf==1.0.6
django-compressor==4.5.1
django-sass-processor==1.4.1
frozenlist==1.5.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
greenlet==3.1.1
gspread==6.1.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
isodate==0.7.2
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
libsass==0.23.0
msal==1.31.1
msal-extensions==1.2.0
multidict==6.1.0
numpy==2.2.1
oauthlib==3.2.2
openai==1.58.1
orjson==3.10.12
outcome==1.3.0.post0
packaging==24.2
pillow==11.1.0
portalocker==2.10.1
propcache==0.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
PyJWT==2.6.0
pyodbc>=5.0.1
PySocks==1.7.1
python-dateutil>=2.8.2
python-dotenv==1.0.1
PyYAML==6.0.2
rcssmin==1.1.2
reportlab==4.2.5
requests
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rjsmin==1.2.2
rsa==4.9
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
sqlparse==0.5.3
tenacity==9.0.0
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.18.3
ffmpeg-python
pytz>=2023.3